requirements.txt
azure-functions
azure-identity
azure-keyvault-secrets
azure-ai-documentintelligence
requests
python-dotenv
azure-storage-blob
